Split sequence to a separate header
