+++ example/batch_inserts_generic.cpp
@@ -25,6 +25,7 @@
 #include <boost/mysql/error_with_diagnostics.hpp>
 #include <boost/mysql/format_sql.hpp>
 #include <boost/mysql/results.hpp>
+#include <boost/mysql/sequence.hpp>
 #include <boost/mysql/string_view.hpp>
 #include <boost/mysql/with_params.hpp>
