+++ example/patch_updates.cpp
@@ -19,6 +19,7 @@
 #include <boost/mysql/error_with_diagnostics.hpp>
 #include <boost/mysql/field_view.hpp>
 #include <boost/mysql/results.hpp>
+#include <boost/mysql/sequence.hpp>
 #include <boost/mysql/string_view.hpp>
 #include <boost/mysql/with_params.hpp>
