+++ include/boost/mysql/detail/format_sql.hpp
@@ -18,10 +18,6 @@
 #include <type_traits>
 #include <utility>
 
-#ifdef BOOST_MYSQL_HAS_CONCEPTS
-#include <concepts>
-#endif
-
 namespace boost {
 namespace mysql {
 
@@ -108,13 +104,6 @@ concept formattable =
     is_formattable_range<T>::value ||
     // This covers passing formattable_ref as a format argument
     is_formattable_ref<T>::value;
-
-template <class FormatFn, class Range>
-concept format_fn_for_range = requires(const FormatFn& format_fn, Range&& range, format_context_base& ctx) {
-    { std::begin(range) != std::end(range) } -> std::convertible_to<bool>;
-    format_fn(*std::begin(range), ctx);
-    std::end(range);
-};
 
 #define BOOST_MYSQL_FORMATTABLE ::boost::mysql::detail::formattable
