+++ test/integration/test/snippets/sql_formatting.cpp
@@ -12,6 +12,7 @@
 #include <boost/mysql/format_sql.hpp>
 #include <boost/mysql/metadata_mode.hpp>
 #include <boost/mysql/results.hpp>
+#include <boost/mysql/sequence.hpp>
 #include <boost/mysql/string_view.hpp>
 #include <boost/mysql/with_params.hpp>
